Fix new course not showing up after creation.
On the list of courses page, if you click Create Course, fill out the
forms and then click Submit, a new course will be added but not shown in
the list until refresh.

This was caused by an exception thrown by an invocation of the
CourseController that was meant to refresh the page.

A refresh of the page shouldn't be necessary, adding the new course to
the $scope.courses variable is sufficient due to 2-way data binding.

The API was modified to return the newly created course's id and name.
This information will then be pushed onto $scope.courses.

To maintain the alphabetical listing of courses, we use angularjs' built
in orderBy filter to sort the courses by course name.