Global courseId not needed in QuestionController.
The courseId does not need to be tracked as a global variable, it should
be passed through the url using the router. Passing the ID in the URL
instead of setting it in Javascript makes it more flexible, e.g.: people
can copy and paste the url to a Questions page directly to another tab
and it would work without having to worry about Javascript state.