Implemented permissions
I have added two permission roles, Teacher and Student. The
permissions checking denies users from doing tasks they are not
supposed to. Also the permissions checking mechanism will throw
an exception if the user is trying to access pages without logging
in. The exception will cause the user to be redirected to the login
page.

When new roles are added, we will need to update the permissions
mechanism.