Added basic installer.
For now all the installer does is redirect to the installer if
tmp/installed.txt does not exist. During the process, the installer
creates an administrator and the installed.txt file. More should be
done for the installer such as the creating the database (asking
for username and password), installing required libraries, and more
system requirement checks on the first page.

Also right now there is a problem when a person refreshes on an
installer view, the user is redirect backed to the first installer
page. This will need to be fixed.