Issue #7: Integrated Gravatar avatar system
I have integrated the Gravatar avatar system into the question and
answer pages (comments included) if the email the user provided
has a valid avatar. If not valid avatar is available a silhouetted
outline of a person is displayed. If we want something more unique
we can used another default image Gravatar provides.

The gravatar hash code is generated from the user model as a vritual
attribute.