Added validation for questions, answers, and comments
I have added simple validations for questions, answers, and comments
to ensure that no empty forms are submitted. If the page only has
one instance of form (eg. question in question page), the error
will appear below the form field similar to create user form. However,
if multiple copies of the form are on the same page (eg. answers),
the error will appear as a flash message. The reason is, it is hard
to keep the toggle open if an error occur and close the toggle
if the submission is successful.

I have also fixed the css for questions and answers comments.