Replace jqueryte with tinymce.
Encountered problems with jqueryte integration with AngularJS where
jqueryte's onchange handler was firing so fast that $scope wasn't
properly updated. This can probably be fixed by using the event system.
But since AngularUI already has a tinymce module, it was easier to just
replace jqueryte with tinymce.

When using ngModel with tinymce, the model stored is actually a tinymce
object. It looks like though, using it as a string seems to work fine,
so I'm guessing there's some sort of translation going on.