Modified create_user and enroll_student to use "import methods"
I have modified create_user and enroll_student to use the import
methods import_users and enrol_users. This is so so that users
will be using the same methods whether they are creating users
and enrolling students via the interface or via import. Due to this
change, the controller also has been slightly modified to accept
different return values for successes and errors.