Minor changes to enrollment filter + more
When you filter a list on ng-Table, it always goes to page 1. (I am assuming it's a correct behavior).
My previous version of enrollment filter didn't make it go to page 1 and there was an issue when user is not on page 1; so I fixed it.
Also, bootswatch.min.css wasn't doing anything other than modifying header sizes, so it is removed.