implemented the math editor toolbar
I have implemented a simple toolbar with sample math formulas to help
users input math formulas into the text editor. The toolbar is created
with two directives, one for the buttons and one for the whole toolbar.

Some improvements can be made such as the amount of time it takes for
the page to load, due to rendering all the math formulas in the
buttons. Another improvement that can be made is to insert the formulas
at the cursor instead of at the end of the text.