Modified Random Judge function
We are limiting the random-judge pool to a 10; meaning, there is a chance of the function not returning a pair even if there is one or more that the user hasn't judged yet. This is no necessarily an error. There was an error, however, in the previous version of random judge; even though there was a fresh pair in pool, it didn't return the pair. Random judge picks 2 fresh pairs (from a randomized, priority-based pool) and compare the sum of the first two lowest count scripts of each pair: lowest one wins. The error is that it REQUIRES 2 pairs; this problem is no fixed by making 3 return cases for the function:
CASE1 - Have 2 or more fresh pairs: compare the first two and return the winner
CASE2 - Have only 1 fresh pair: return the first pair
CASE3 - No fresh pair at all: return nothing (again, not necessarily an error)