Implement select toggle for mathtoolbar
I have implemented a drop down menu to toggle through the various
math toolbars. The initial state is to not show any math toolbars
at all, so that the user will not be overwhelmed.

I have also fixed the problem with creating an admin during
installation. The reason was one of the course methods had a
teacher requirement, but during installation the user has no roles
at all.

I have also fixed Issue 25 where during installation the admin's
email is not saved. The reason was a logic problem when validating
the email field. It was slightly different from the logic used in
create user (correct one).