Many updates including faster page rendering for mathjax
- update server (Flask) to check for installation status (existence
of installed.txt). The server checks now for the file's existence
instead of whether the file is writable.

- updated the Question model's title field to allow up to 255
characters instead of 80, which is too short.

- fixed answer page's question title to render math equation

- fixed update answer comments' edit submit button to use bootstrap
classes

- changed enroll page's status syntax. Instead of saying the student
is dropped, it now says unenrolled. More upgrades should be done to
the enrol/unenrol toggle to make it more obvious. Currently, it looks
like it is only plain text.

- sped up question and answer page rendering by not loading the
math toolbars on first load. The toolbars will only be rendered
when they are selected in the drop down menus.