Convert 'flash' and 'installed' provider to module
Since flash and installed providers are used by other modules across the
application, they are placed into the modules/common/ directory.

This commit is meant as an example of a trivial conversion of existing
components to modules. Note that these new modules haven't actually been
hooked up, that will be done in the next commit.