Refactor top navbar, implement logout.
The top navbar needed to be refactored for the new login system. It now
uses the AuthenticationService to check user login information.

The IndexController has been renamed NavbarController and is now in its
own module.

Removed tutorial button from the top navbar. I think the whole tutorial
option should be triggered on individual module pages, instead of being
in the navbar. This should make it easier to implement, as we don't have
to worry about communication between different controllers now.

The login module now has a logout controller. The logout controller is
only used for the logout option in the top navbar's user dropdown. Not
sure if this is the best organizational arrangement, but it'll have to
do.

The navbar is now 100% stretched across the browser instead of being
confined to the rest of the container. It should be fluid now and should
shrink to a collapsible button on small screens.