Login now shows as a modal popup.
On auth failure, the user is redirected to the login page. This means
that the user loses all the context of the page they were on.
Http-auth-interceptor suggests that no redirection is performed on the
page in order for it's failed request replay to work properly.
Implementing the login as a modal allows us to ask the user for
authentication without losing the context of where they were before
going to the login page. The modal shows up as an overlay, the user
enters their information, the failed requests are automatically
repeated, and the user should see everything resume seamlessly.

This means that only pages that need to make authenticated requests will
see the login popup. So if we need to make pages that don't need
authentication, all we need to do is make sure no authentication
requried requests are made on that page.

The http-auth-interceptor library was updated to 1.2.1 in this commit.