Fix autofocus for login form.
The switch to modal for the login form broke auto focus on the Username
input. Looks like the HTML5 attribute autofocus doesn't work either.
Both problems seems to be caused by not waiting for the input to render
before setting focus.

Using a $timeout looks like is the standard way for forcing a wait for
render, that seems to fix the auto focus directive.

Since the auto focus directive seems to be only used in the login form,
I've moved it into the login module. If it's used elsewhere, then we'll
need to look at putting it somewhere that makes more sense.