Home page refactor.
The home page was moved from coursepage to the Angular module "home".

Also tried to get Flask-Restless to play well with not exposing the
user's password hash. Unfortunately, all the different ways of excluding
fields or only including fields seem to fail when doing queries on
relations (e.g.: /users/1/courses). The password hash is hidden on
normal users queries but when you ask for something like a user's
courses, it pops back up. This means that we might have to abandon
Flask-Restless.