Course API updated, new add course page.
The Course API is now switched to the Flask-Restless assisted module.
Moved the to_dict() function out of User to Util, since it's now also
used by Course. There is now a pagination variant called
to_dict_paginated(). In addition to doing the actual pagination, it also
fills in the fields necessary for pagination support on the client side.

Adding a new course is now in it's own page instead of being on the home
page. It now lives in the course module. We should be able to achieve
the same effect of not leaving the home page by putting it in a modal
like the login. There's problems with the login modal though where the
home page doesn't update after a login, need to solve that first.

While permissions on the server side is nicely handled by Flask-Bouncer,
there's no equivalent on the client side. This will need to be addressed.