New breadcrumbs mechanism.
Breadcrumbs are now specified at $routes specification rather than a
value set by each controller. We're using the ng-breadcrumbs library
(https://github.com/ianwalter/ng-breadcrumbs).

When configuring routes, a "label" attribute is specified for the route,
this is used as part of the breadcrumb. The breadcrumb appears to be
formed by unravelling the url path.