Permissions retrieved on login, fix login trigger
Performing an extra request to retrieve authorization information turns
out be to be problematic due to having to Angular's insistence on async
requests. It was easier to just integrate it into login. So now when
users login, we'll get the permissions information for that user too.

The LoginService will give the permissions info to the Authorize
service.

Also, the login box doesn't seem to trigger reliably when not logged in
and on logout. This turned out to be due to an erroneous assumption on
my part that the http-auth-interceptor library could handle it all. Some
thinking made me realize that http-auth-interceptor actually only
handles one case of the login (server logouts while browser session
still valid). So the "new user who hasn't logged in yet" and "user who has
just logged out" cases weren't handled.

Provisions were added in the login-module to handle those cases, mainly
modifications to the module initialization to check for user
authentication on every route change and shows the login box if user
isn't authenticated.