Use CKEditor for textarea, new home course listing
Since CKEditor has a nice Mathjax integration, we might as well take
advantage of it instead of having to create our own math toolbar in
hallo.js. First try at using it in the Create Course form with a very
simplified toolbar set for the course description.

Not sure if getting rid of the table styled course listing is the best
idea, but trying it out. It displays the course name and the course
description. Clicking on it should go to the course in question, but
that hasn't been implemented yet. There should also be a count of
questions and comments that are new since the last visit.