Fix login test case.
Added permissions info to login, so needed to change the test case.
Encountered problem trying to get the logged in user's application
context so that Flask-Bouncer and Flask-Login's helper functions can be
used. Couldn't solve it, so as a workaround, I'm just manually logging
the user in with Flask-Login's login_user() function in whatever
application context the test case happens to be running in at the time.