Replace flash messages with AngularJS-Toaster.
Notifications are now provided by AngularJS-Toaster. There were problems
with the two Angular flash messaging libraries I tried out: their divs
take up space while empty/too limited in styling configurations/doesn't
allow html in the messages. Toaster seems to solve all these problems in
that it doesn't take up space while empty, good overrides for
controlling styles (though the default is good enough that I didn't
bother), and it has options for allowing html.

Note that I've extended the toaster provider with success(), error(),
etc methods so that we don't have to provide all the params for pop().
There is also reqerror() that is meant to provide a generic message for
all ajax request errors.

Not sure if this is the best way to have error templates, but going with
it for now.