Restrict information sent about users.
User contributions under the cover of anonymity tends to be more free
flowing with the downside that it can also be more hostile. We'd like
the benefit of the former without encouraging the latter. So we're going
to allow students to conceal identity from other students, to encourage
contribution. But we're also providing instructors access to the real
identities, for accountibility.

Students can only see the user's chosen "displayname" while instructors
have access to the user's real name.

If a user has permission to edit a user's enrolment, then we assume that
they have permission to view other user's real names.