Convert Courses API to use Flask-Restful.
This is the last API to be converted. The Flask-Restless based helper
code has also been removed.

Flask-Restful requires you to specify the fields that are transmitted to
the user. This now lives in the dataformat module. Would've named it the
fields module except then we'd have a confusion with Flask-Restful's
fields module.