Single form for editing/creating courses.
Instead of duplicating code and having separate forms for editing or
creating courses, there is now a single form that should serve both
purposes. There is some trade off, such as not being able to tailor the
submit button for Edit/Create courses, leaving it as a generic "Save"
button.

Putting all the css and elements required to implement bootstrap's form
field feedback look is now abstrated into a directive, the
acj-field-with-feedback directive. It's first used is in the course
form.

CKEditor options are now centralized into the EditorOptions service.
Though currently there is only the "basic" function.