Flask-Bouncer throws 401, we want 403.
Flask-Bouncer throws 401 Authentication Required if permission checking
fails. We want it to throw 403 Forbidden instead. The 401 triggers the
login box on the UI side, even though login wouldn't help the user since
they don't have permission in the first palce. 403 is the more accurate
error to throw.

Instead of modifying the library, the Authorization module now has a
require() function that wraps Flask-Bouncer's ensure(). It'll throw a
403 Forbidden instead of a 401. Note that this means Flask-Bouncer's
requires() decoration shouldn't be used either and should just be a
require() call.