Remove setting TESTING:True from test, minor fixes
The Flask app config variable "TESTING", when set to True, causes
Flask-Login to ignore login_required decorators. This means that the
unauthorized test cases were passing due to Flask-Bouncer throwing a
401, not because Flask-Login throws the 401. Now that we throw 403 on
Flask-Bouncer errors, the test cases broke. I've removed the TESTING
variable but not sure if it's required for other testing functionality,
so I've kept it commented with an explanation for now.

Fixed typo in question-create-partial.html.

Added missing imports in authorization.py.