Added test cases for getting and creating comments
I've found a bug with some of the test cases where the expected
404 is returned but not for the reason we expected. It was
returning 404 for an invalid url instead of the expeced invalid
id. The appropriate changes have been made. Additionally, some
queries are added to check for the ids' validity (until a better
solution is found)

I've also added permissions for comments for the different user
types.