Modified creating/editing comments interface
Firstly, I fixed the answer model that was trying to query Questions,
which doesn't exist. Instead it is PostsForQuestions.

I have also modified the comments form to display the question and
answer the comment relates to. Also I have modified the controller
to redirects to an appropriate page after editing a question and a
comment