Implemented user creation
I've created a form to manually create users. A link has also been
added to the navigation bar (as long as the user has the permission
to create users). A new directive has been created to check the
password and "verifyPassword" matches. Furthermore, the username
and display name is checked to be unique before the user is created.