Add test cases for Judgements API
Get answer pair has a 'second' half of tests in the submit judgement
section. The second half of tests is meant to verify behaviour when
there's already been judgements made on some answers.

Submit judgement now returns the judgements stored to make it more
inline with the other APIs. This also makes the testing faster as the
data to be verified can be seen in the return.