Implemented importing class enrolment
Currently, we are using the angular-file-upload since angularjs does
not have a native file upload module. More information can be found
at https://github.com/nervgh/angular-file-upload. It has been setup
to be installed with bower

The current implementation only allows csv formatted class lists and
only requires usernames of the users. If no passwords (optional) are
given for the users, their username would be used for their password.
If a given display name has been taken, then we will leave the field
empty to allow the user to enter another display name later. I've setup
an import service to so that the code can be reused and allow sharing
of variables between the initial import page and results page.
Currently, the upload files are saved in tmpUpload/.

The link to the import page is found in the navigation bar if logged
in user has permissions to create users and are in a course. After a
successfuly upload, the user would be redirected to a results page.

For this implementation to work, the database has to be recreated to
get the modification I made to the Users table. Furthermore, bower
would have to be ran again to install the required modules.