merged user profile with users view
Now the interface for user profile is the same as users view. The
only difference is the system will recognize the logged in user
is the same as the viewed user, therefore an edit option will be
available.

The edit profile interface is the same as edit user's interface with
the addition of the option to change the password.

The password match directive has been added so that the submit
button will not be active until the two password fields in the form
have identical values.

The css file will have to be regenerated from the less files to
receive the new css class I created for the users view interface