Implement populate with sample data to manage.py
Add the -s flag when creating or recreating the database to add sample
data, e.g.: `python manage.py database recreate -s`

All this does is add 5 courses, an instructor, and 6 students. The
instructor is named "instructor1" while the students are named
"student1","student2", and so on.

Each student and instructor is enroled in 3 courses, which has 5
questions and corresponding answers from students.

Although it's called "sample" data, it's really more like testing data
right now, since it's just randomly generated non-sense, most of it.