implemented ranking mechanism (generate scores)
The ranking mechanism is used when a certain criteria is met.
Currently, the condition is each answer must be judged at least
once. During the process, three values are generated which are:

rounds - number of times the answer has been compared to another
         answer
wins - number of times the answer has won against another answer
score - the ACJ algorithm computed score. For more details refer
	to the research paper

Currently the score is printed (rounded to three decimal places)
next to the answer's post date. The score used is also the default
criteria.

In the future, when we have implemented the ability to have multiple
criteria, we'll need a way to allow users to sort the answers by
the preferred criterion.

Notes:
- Still need to sort the answers in order of their score.
- When we increase the python version for the application to be run
on, we'll need to update some of the code that uses itertools in
models.py. Eg. change ifilterfalse to filterfalse
- also the database needs to be recreated for this implementation to
work. I've updated the Scores table.