ranked the answers by their scores
The question api has been modified to return not only the question
but also the list of related criteria (criteriandcourses) and also
the number of enroled students. The answers are also returned in
descending order by their posted date. However, if scores are
available the answers will be reordered in descending order by
their score

A dropdown menu has been created to allow users to change the order
by changing to their preferred critera. However, currently we only
have the one basic criterion and not have the ability to add more,
so hid the dropdown menu from the users to avoid confusion. The scores
has also been displayed in a format to allow it to change when the
crierion change in the dropdown menu.

Note:
- a minor bug has been noticed. when some answers do not have scores
but others do, the ones that do not have scores will appear on the top.
This will have to be fixed. This bug will not appear if we have separate
answering and judging periods.

Tests have also been modified to reflect the changes made to the questions
api.