Add ability to flag answers on Judgement page.
Answers can now be flagged as inappropriate or incomplete on the
Judgement page.

Anybody can flag an answer, but only the person who flagged it or TA's
and above can unflag an answer.

There should probably be a visual indicator for instructors to know that
certain answers has been flagged. There should probably also be the flag
button on the view question page too.

The person who flagged the answer is stored, so instructors can know who
flagged what answer.