Implemented "required N judgements" counter on course home page
I have implemented a "counter" for number of required judgements
left for the user to complete. It is currently labeled as
"Judgements Left".

The required number of judgements per user is calculated by:

x = required # of rounds per pair = 6
y = minimum # of pairs of answers  = # of answers / 2
required # of judgements = (x*y) / # of students