Refactor authentication and authorization module, add Session service
Now Session service as a data model is providing user and permission
information. Both getUser and getPermissions will return `thenable`
promise and the resolved value is the user/permission object. The
controller can hold a reference of user object for view. The getUser
will always return the same object so the reference in controller will
always be valid.