implemented answer deletion
I've implemented the ability to delete answers as long as the user
can manage the course or the author of the answer. I've also modified
question deletion. I removed the extra question get request, since it
is done in delete api already. I've also fixed the bug where a question
cannot be deleted if there are any answers or comments relating to it.
The reason is the deletion is not cascading. Tests have also been
added to test the delete api