Fix timestamps defaulting to local timezone.
The defaults for 'created' and 'modified' defaults and updates to
current timestamp, which is the time in the local timezone. This
conflicts with our assumption that all stored time in the database is
UTC and makes the UI display wrong times when the defaults are used.

Fix is to use datetime.datetime.utcnow instead of current_timestamp.
Note that this fix is only valid if data is inserted through SQLAlchemy.
Raw queries to the database will no longer has defaults since
'datetime.datetime.utcnow' is a Python function and not a SQL function.
This was necessary because even though there is a utc counterpart to
current_timestamp called utc_timestamp in MySQL, Sqlite doesn't support
utc_timestamp, or at least, SQLAlchemy doesn't seem to translate it
properly causing an exception.

Note that the ability to have auto-filling columns for modified was why
I went with Timestamp type columns in the first place for the 'created'
and 'modified' fields. Now that the auto-filling isn't possible, I've
switched the data type from Timestamp to DateTime. AFAIK, DateTime
should prove to be more robust and convenient in terms of
future-proofing.