Fix judgment submit working only once sometimes.
When judging a question, if another question has answer pairs in a
higher round, then judgment submit will only work once. This is due to
the query selecting the max rounds first and then applying the question
restriction. So the rounds number we get is the max round out of every
single question, instead of just the question we're looking at.

Fix is to split the max round and question restrictions into subqueries
and make sure that the question restriction one is executed first.