Another overhaul of the answer pairing system.
This one gets rid of the rounds based system, which should improve
answer pair availability, this should fix #77. It is a bit more
processing intensive though, as this implementation goes back to the on
demand generation system. It shouldn't be too difficult to change back
to pre-generated answer pairs if necessary though.

The basic idea behind this overhaul is that answers which are scored
close to each other should get more judgements than answers which are
outliers from the rest of the other answers.

The AnswerPairGenerator class now contains all the answerpair generation
code. Like the previous system, it randomly pairs answers that hasn't
been judged yet first. Unlike the previous system, it'll be tolerant of
new answers coming in after judgements have been submitted.