First commit for adding new criteria to courses
- modified criteria api to have two paths, one for CriteriaAndCourses
and the other for just Criteria. This allow us the flexibility of
only interacting with the Criteria or the relationship it has with
courses
	- CoursesCriteriaResource - for model CoursesAndCriteria
	- CriteriaResource - for model Criteria
- current design is to have the add criteria option part of the
course configure form. When a new criteria is created, it is by default
added to the course as well
- editing criteria is only available to the author and admins. The
edit form is a new screen. When the update is submitted, the user
is redirected back to the course configure form they came from
- the add and edit form uses the same partial
- a partial has also been partially created for listing the criteria
currently being attached to the course. This partial is still under
construction
	- adding remove action
	- dropdown menu to add previously created courses
	- also intended for use in the question form