Fixed Issue #80
The permissions are generated at login, therefore if the instructor
does not have any courses, they would not have permissiont to create
questions for example. Furthermore, the permissions are not being
updated after course creation, which is why the buttons are unavailable.

To fix this, I created a refresh function for the Session service,
which updates the permissions (including the new course).

On a side note, I have fixed the api call "/session/permissions" (typo)