Fix answer pairing availability, same answer pair
If no unscored pairs can be formed, then we have to return a scored
pair. Previously, it was just failing.

When forming unscored answer pairs, there was a possibility that it can
return a pair where both answers are the same due to randomly picking
from a list of all available answers. It should now pick from a list of
all available answers but minus the currently unscored answer.